float to double

+++ EnergyMinimizer.hpp
@@ -6,7 +6,7 @@ typedef boost::shared_array<short> short_array;
 class EnergyMinimizer
 {
     public:
-   virtual short_array minimize(short_array input, float lambda, float& energy, float &m, float& b) = 0;
+   virtual short_array minimize(short_array input, double lambda, double& energy, double &m, double& b) = 0;
 
    virtual ~EnergyMinimizer(){}
 };
